Run a single socsim-simulation with a given supplement-file and folder. The results will be saved into that folder

Run a single socsim-simulation with a given supplement-file and folder. The results will be saved into that folder

run1simulationwithfile(
  folder,
  supfile,
  seed = "42",
  process_method = "inprocess"
)

Arguments

folder

base-directory of the simulation.

supfile

the .sup file to start the simulation, relative to the folder

seed

RNG seed as string, Default="42"

process_method

specify whether and how SOCSIM should be started in its own process or in the running R process. Use one of "inprocess" - SOCSIM runs in the R-process. Beware if you run several different simulations - they may affect later simulations "future" - the safest option. A new process will be start via the "future" package "clustercall" - if the future package is not available, try this method instead

rootfolder

rootfolder...

Value

The results will be written into the specified folder